requirements.txt
python-telegram-bot==20.7
anthropic==0.40.0
